dashboard/requirements.txt
flask==3.1.0
flask-socketio==5.4.1
requests==2.32.3
python-dotenv==1.0.1
python-socketio[client]==5.12.1
gunicorn==21.2.0
eventlet==0.36.1
